requirements.txt
flask
aiohttp
python-dotenv
botbuilder-core
botbuilder-schema
snowflake-snowpark-python
snowflake-connector-python
cryptography
